telegram-sales-bot/requirements.txt
# Core dependencies
python-telegram-bot==21.3
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.31
alembic==1.13.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
jinja2==3.1.4
aiofiles==24.1.0

# Payment gateways
mercadopago==2.2.1
stripe==10.5.0

# Database
asyncpg==0.29.0
psycopg2-binary==2.9.9
aiosqlite==0.20.0

# Redis for caching
redis==5.0.7
aioredis==2.0.1

# Environment and config
python-dotenv==1.0.1
pydantic==2.8.2
pydantic-settings==2.4.0

# HTTP client
httpx==0.27.0
aiohttp==3.10.5

# Utilities
python-dateutil==2.9.0
pytz==2024.1
qrcode[pil]==7.4.2
Pillow==10.4.0

# Logging
loguru==0.7.2

# Testing
pytest==8.3.2
pytest-asyncio==0.23.8
pytest-cov==5.0.0
